+++ app/views/layouts/application.html.erb
@@ -9,8 +9,6 @@
 
     <%= stylesheet_link_tag "application", "data-turbo-track": "reload" %>
     <%= javascript_importmap_tags %>
-    <%= javascript_import_module_tag "simurate" %>
-    <%= javascript_import_module_tag "simulation" %>
   </head>
 
   <body>
